+++ packages/manager/apps/hub/src/dashboard/incident/status/status.component.js
@@ -1,0 +1,12 @@
+import controller from './status.controller';
+import template from './template.html';
+
+export default {
+  bindings: {
+    services: '<',
+    translatedServiceEnum: '<',
+    translatedStatusEnum: '<',
+  },
+  controller,
+  template,
+};