+++ packages/manager/apps/hub/src/dashboard/incident/status/status.controller.js
@@ -1,0 +1,19 @@
+import { Environment } from '@ovh-ux/manager-config';
+import { FAQ } from './status.constants';
+
+export default class StatusController {
+  $onInit() {
+    this.filtersOptions = {
+      status: {
+        hideOperators: true,
+        values: this.translatedStatusEnum,
+      },
+      products: {
+        hideOperators: true,
+        values: this.translatedServiceEnum,
+      },
+    };
+
+    this.faqLink = FAQ[Environment.getUserLocale()] || FAQ.en_GB;
+  }
+}