+++ packages/manager/apps/hub/CHANGELOG.md
@@ -1,3 +1,30 @@
+# [3.5.0](https://github.com/ovh/manager/compare/@ovh-ux/manager-hub-app@3.4.0...@ovh-ux/manager-hub-app@3.5.0) (2021-03-15)
+
+
+### Features
+
+* **dasboard:** add banner incident for non FR customer ([#4548](https://github.com/ovh/manager/issues/4548)) ([11da491](https://github.com/ovh/manager/commit/11da4913a531381c2c10f39aeeb021c7e2799302))
+
+
+
+# [3.4.0](https://github.com/ovh/manager/compare/@ovh-ux/manager-hub-app@3.3.0...@ovh-ux/manager-hub-app@3.4.0) (2021-03-15)
+
+
+### Features
+
+* **hub:** add service url in incident status list ([#4546](https://github.com/ovh/manager/issues/4546)) ([58f1010](https://github.com/ovh/manager/commit/58f1010dce715180bea589228c1bc85143be20ce))
+
+
+
+# [3.3.0](https://github.com/ovh/manager/compare/@ovh-ux/manager-hub-app@3.2.6...@ovh-ux/manager-hub-app@3.3.0) (2021-03-14)
+
+
+### Features
+
+* add incident status listing ([#4543](https://github.com/ovh/manager/issues/4543)) ([b676514](https://github.com/ovh/manager/commit/b676514271f0498464741752107cfdfb24d82e3b))
+
+
+
 ## [3.2.6](https://github.com/ovh/manager/compare/@ovh-ux/manager-hub-app@3.2.5...@ovh-ux/manager-hub-app@3.2.6) (2021-03-11)